+++ dpnp/__init__.py
@@ -304,6 +304,7 @@ from .dpnp_iface_indexing import (
     unravel_index,
 )
 from .dpnp_flatiter import flatiter
+from .dpnp_broadcast import broadcast
 
 # -----------------------------------------------------------------------------
 # Linear algebra
@@ -691,6 +692,7 @@ __all__ += [
     "atleast_1d",
     "atleast_2d",
     "atleast_3d",
+    "broadcast",
     "broadcast_arrays",
     "broadcast_to",
     "column_stack",